ai_agents/sales_variables_extractor/requirements.txt
# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# LangChain and LangGraph
langchain>=0.1.0
langchain-google-genai>=0.0.6
langgraph>=0.0.20

# Google Generative AI (New package)
google-genai>=0.2.0

# Pydantic for data validation
pydantic>=2.0.0

# Optional: LangSmith for tracing
langsmith>=0.0.70

